requirements.txt
pyside6
peewee
pillow